third_party/blink/web_tests/external/wpt/webnn/idlharness.https.any.worker-expected.txt

This is a testharness.js-based test.
Found 46 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] idl_test setup
  promise_test: Unhandled rejection with value: object "NotSupportedError: The input layout nchw is not supported."
[FAIL] NavigatorML must be primary interface of navigator
  assert_own_property: self does not have own property "NavigatorML" expected property "NavigatorML" missing
[FAIL] Stringification of navigator
  assert_class_string: class string of navigator expected "[object NavigatorML]" but got "[object WorkerNavigator]"
[FAIL] NavigatorML interface: navigator must not have property "ml"
  assert_false: expected false got true
[FAIL] ML interface: operation createContextSync(optional MLContextOptions)
  assert_own_property: interface prototype object missing non-static operation expected property "createContextSync" missing
[FAIL] ML interface: operation createContextSync(GPUDevice)
  assert_own_property: interface prototype object missing non-static operation expected property "createContextSync" missing
[FAIL] ML interface: navigator.ml must inherit property "createContextSync(optional MLContextOptions)" with the proper type
  assert_inherits: property "createContextSync" not found in prototype chain
[FAIL] ML interface: calling createContextSync(optional MLContextOptions) on navigator.ml with too few arguments must throw TypeError
  assert_inherits: property "createContextSync" not found in prototype chain
[FAIL] ML interface: navigator.ml must inherit property "createContextSync(GPUDevice)" with the proper type
  assert_inherits: property "createContextSync" not found in prototype chain
[FAIL] ML interface: calling createContextSync(GPUDevice) on navigator.ml with too few arguments must throw TypeError
  assert_inherits: property "createContextSync" not found in prototype chain
[FAIL] MLCommandEncoder interface: existence and properties of interface object
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLCommandEncoder interface object length
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLCommandEncoder interface object name
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLCommandEncoder interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLCommandEncoder interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLCommandEncoder interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLCommandEncoder interface: operation initializeGraph(MLGraph)
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLCommandEncoder interface: operation dispatch(MLGraph, MLNamedGPUResources, MLNamedGPUResources)
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLCommandEncoder interface: operation finish(optional GPUCommandBufferDescriptor)
  assert_own_property: self does not have own property "MLCommandEncoder" expected property "MLCommandEncoder" missing
[FAIL] MLContext interface: operation computeSync(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews)
  assert_own_property: interface prototype object missing non-static operation expected property "computeSync" missing
[FAIL] MLContext interface: operation createCommandEncoder()
  assert_own_property: interface prototype object missing non-static operation expected property "createCommandEncoder" missing
[FAIL] MLContext interface: context must inherit property "computeSync(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews)" with the proper type
  assert_inherits: property "computeSync" not found in prototype chain
[FAIL] MLContext interface: calling computeSync(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews) on context with too few arguments must throw TypeError
  assert_inherits: property "computeSync" not found in prototype chain
[FAIL] MLContext interface: context must inherit property "createCommandEncoder()" with the proper type
  assert_inherits: property "createCommandEncoder" not found in prototype chain
[FAIL] MLGraph must be primary interface of graph
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: graph is not defined"
[FAIL] Stringification of graph
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: graph is not defined"
[FAIL] MLGraphBuilder interface: operation constant(MLOperandDescriptor, MLBufferView)
  assert_equals: property has wrong .length expected 1 but got 2
[FAIL] MLGraphBuilder interface: operation constant(double, optional MLOperandDataType)
  assert_equals: property has wrong .length expected 1 but got 2
[FAIL] MLGraphBuilder interface: operation buildSync(MLNamedOperands)
  assert_own_property: interface prototype object missing non-static operation expected property "buildSync" missing
[FAIL] MLGraphBuilder interface: operation not(MLOperand)
  assert_own_property: interface prototype object missing non-static operation expected property "not" missing
[FAIL] MLGraphBuilder interface: operation gru(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLGruOptions)
  assert_own_property: interface prototype object missing non-static operation expected property "gru" missing
[FAIL] MLGraphBuilder interface: operation gruCell(MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLGruCellOptions)
  assert_own_property: interface prototype object missing non-static operation expected property "gruCell" missing
[FAIL] MLGraphBuilder interface: operation lstmCell(MLOperand, MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLLstmCellOptions)
  assert_own_property: interface prototype object missing non-static operation expected property "lstmCell" missing
[FAIL] MLGraphBuilder interface: operation triangular(MLOperand, optional MLTriangularOptions)
  assert_own_property: interface prototype object missing non-static operation expected property "triangular" missing
[FAIL] MLGraphBuilder interface: builder must inherit property "buildSync(MLNamedOperands)" with the proper type
  assert_inherits: property "buildSync" not found in prototype chain
[FAIL] MLGraphBuilder interface: calling buildSync(MLNamedOperands) on builder with too few arguments must throw TypeError
  assert_inherits: property "buildSync" not found in prototype chain
[FAIL] MLGraphBuilder interface: builder must inherit property "not(MLOperand)" with the proper type
  assert_inherits: property "not" not found in prototype chain
[FAIL] MLGraphBuilder interface: calling not(MLOperand) on builder with too few arguments must throw TypeError
  assert_inherits: property "not" not found in prototype chain
[FAIL] MLGraphBuilder interface: builder must inherit property "gru(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLGruOptions)" with the proper type
  assert_inherits: property "gru" not found in prototype chain
[FAIL] MLGraphBuilder interface: calling gru(MLOperand, MLOperand, MLOperand, unsigned long, unsigned long, optional MLGruOptions) on builder with too few arguments must throw TypeError
  assert_inherits: property "gru" not found in prototype chain
[FAIL] MLGraphBuilder interface: builder must inherit property "gruCell(MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLGruCellOptions)" with the proper type
  assert_inherits: property "gruCell" not found in prototype chain
[FAIL] MLGraphBuilder interface: calling gruCell(MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLGruCellOptions) on builder with too few arguments must throw TypeError
  assert_inherits: property "gruCell" not found in prototype chain
[FAIL] MLGraphBuilder interface: builder must inherit property "lstmCell(MLOperand, MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLLstmCellOptions)" with the proper type
  assert_inherits: property "lstmCell" not found in prototype chain
[FAIL] MLGraphBuilder interface: calling lstmCell(MLOperand, MLOperand, MLOperand, MLOperand, MLOperand, unsigned long, optional MLLstmCellOptions) on builder with too few arguments must throw TypeError
  assert_inherits: property "lstmCell" not found in prototype chain
[FAIL] MLGraphBuilder interface: builder must inherit property "triangular(MLOperand, optional MLTriangularOptions)" with the proper type
  assert_inherits: property "triangular" not found in prototype chain
[FAIL] MLGraphBuilder interface: calling triangular(MLOperand, optional MLTriangularOptions) on builder with too few arguments must throw TypeError
  assert_inherits: property "triangular" not found in prototype chain
Harness: the test ran to completion.
